Move mock_robot.py -> mockrobot.py
